Added two checks of four checks for data
I added checking of the achievement id and the app key, and I wrote some
of the checking for the username, but commented it out to test the
achievement id checking.

I added an app_name field to the achievement update, and changed it so
that appname isn't in the url. This is so that I can use the app name
while validating the data, without having to pass it around.

I added some type declarations to functions in main that were missing it
for no reason.

I added a bunch of test files that test every checking codepath I've
written so far. I will add more as I write more checks.

I added a bunch more functions for getting data from the achievements
database to check against.