Fixed a pile of warnings, added more checks
I fixed all of the redunant imports and all of the type signature
missing warnings that I can right now. I can't currently fix the ones in
dbtalk that include Convertible in the type signature because they need
specific flags set in ghc to compile at all, and I want to look into
them more before I turn them on, because it currently works fine just
ignoring the warnings for those.

I updated the check for the id, but I need to rework it a bit, so the
old one is still in use, and the new one is commented out.

I wrote the check progress check, the first part needs to be fixed so
that it will deal with achievements that are updated without a progress
provided that already have current progress in the database, by
subtracting the current progress from the max progress. That will be in
the next commit shortly.